.history/requirements_20250701121200.txt
# requirements.txt şöyle görünmeli:
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
google-auth>=2.23
google-auth-oauthlib>=1.1
google-api-python-client>=2.100